refactor: Crawler api refactoring
