+++ src/main/java/yagu/yagu/game/GameScheduleCrawler.java
@@ -1,4 +1,4 @@
-package yagu.yagu.crawler;
+package yagu.yagu.game;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
@@ -10,13 +10,13 @@ import org.openqa.selenium.support.ui.Select;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
+import yagu.yagu.game.entity.KboGame;
+import yagu.yagu.game.repository.KboGameRepository;
 
 import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
