+++ src/main/java/yagu/yagu/game/entity/KboGame.java
@@ -1,9 +1,10 @@
-package yagu.yagu.crawler;
+package yagu.yagu.game.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import yagu.yagu.game.Status;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
