+++ src/main/java/yagu/yagu/game/repository/KboGameRepository.java
@@ -1,7 +1,7 @@
-package yagu.yagu.crawler;
+package yagu.yagu.game.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import yagu.yagu.crawler.KboGame;
+import yagu.yagu.game.entity.KboGame;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -12,6 +12,4 @@ public interface KboGameRepository extends JpaRepository<KboGame, Long> {
     Optional<KboGame> findByGameDateAndGameTimeAndHomeTeamAndAwayTeam(
             LocalDate date, LocalTime time, String homeTeam, String awayTeam);
 
-    List<KboGame> findByHomeTeamContainingOrAwayTeamContainingOrderByGameDateAscGameTimeAsc(
-            String home, String away);
 }
